also move is*Key-guard-functions into shared
